r30|initial version

+++ trunk/src/tk/treectrl/qebind.c
@@ -1,7 +1,7 @@
 // ======================================================================
 // Author : $Author$
-// Version: $Revision: 28 $
-// Date   : $Date: 2011-05-21 14:57:26 +0000 (Sat, 21 May 2011) $
+// Version: $Revision: 30 $
+// Date   : $Date: 2011-05-23 14:49:04 +0000 (Mon, 23 May 2011) $
 // Url    : $URL$
 // ======================================================================
 
@@ -36,7 +36,6 @@
 
 #include <ctype.h>
 #include <string.h>
-#include <stdint.h>
 #include <tcl.h>
 #include <tk.h>
 #include "qebind.h"
@@ -232,7 +231,7 @@ int QE_InstallEvent(QE_BindingTable bindingTable, char *name, QE_ExpandProc expa
 
 	Tcl_SetHashValue(hPtr, (ClientData) eiPtr);
 
-	hPtr = Tcl_CreateHashEntry(&bindPtr->eventTableByType, (char *)(intptr_t) type, &isNew);
+	hPtr = Tcl_CreateHashEntry(&bindPtr->eventTableByType, (char *)(long) type, &isNew);
 	Tcl_SetHashValue(hPtr, (ClientData) eiPtr);
 
 	/* List of EventInfos */
@@ -356,7 +355,7 @@ int QE_UninstallEvent(QE_BindingTable bindingTable, int eventType)
 	int i, count = 0;
 
 	/* Find the event */
-	hPtr = Tcl_FindHashEntry(&bindPtr->eventTableByType, (char *)(intptr_t) eventType);
+	hPtr = Tcl_FindHashEntry(&bindPtr->eventTableByType, (char *)(long) eventType);
 	if (hPtr == NULL)
 		return TCL_ERROR;
 	eiPtr = (EventInfo *) Tcl_GetHashValue(hPtr);
@@ -463,7 +462,7 @@ static EventInfo *FindEvent(BindingTable *bindPtr, int eventType)
 {
 	Tcl_HashEntry *hPtr;
 
-	hPtr = Tcl_FindHashEntry(&bindPtr->eventTableByType, (char *)(intptr_t) eventType);
+	hPtr = Tcl_FindHashEntry(&bindPtr->eventTableByType, (char *)(long) eventType);
 	if (hPtr == NULL) return NULL;
 	return (EventInfo *) Tcl_GetHashValue(hPtr);
 }
