+++ trunk/src/tk/treectrl/tkTreeColumn.c
@@ -1,7 +1,7 @@
 // ======================================================================
 // Author : $Author$
-// Version: $Revision: 28 $
-// Date   : $Date: 2011-05-21 14:57:26 +0000 (Sat, 21 May 2011) $
+// Version: $Revision: 30 $
+// Date   : $Date: 2011-05-23 14:49:04 +0000 (Mon, 23 May 2011) $
 // Url    : $URL$
 // ======================================================================
 
@@ -18,7 +18,6 @@
  * RCS: @(#) $Id: tkTreeColumn.c,v 1.87 2009/05/17 19:46:27 treectrl Exp $
  */
 
-#include <stdint.h>
 #include "tkTreeCtrl.h"
 
 typedef struct TreeColumn_ TreeColumn_;
@@ -467,7 +466,7 @@ ColumnCO_Set(
 	int flags					/* Flags for the option, set Tk_SetOptions. */
 	)
 {
-	int cfoFlags = (int)(intptr_t) clientData;
+	int cfoFlags = (long) clientData;
 	TreeCtrl *tree = (TreeCtrl *) ((TkWindow *) tkwin)->instanceData;
 	int objEmpty;
 	TreeColumn new, *internalPtr;
