+++ trunk/src/tk/treectrl/tkTreeCtrl.c
@@ -1,7 +1,7 @@
 // ======================================================================
 // Author : $Author$
-// Version: $Revision: 28 $
-// Date   : $Date: 2011-05-21 14:57:26 +0000 (Sat, 21 May 2011) $
+// Version: $Revision: 30 $
+// Date   : $Date: 2011-05-23 14:49:04 +0000 (Mon, 23 May 2011) $
 // Url    : $URL$
 // ======================================================================
 
@@ -18,7 +18,6 @@
  * RCS: @(#) $Id: tkTreeCtrl.c,v 1.110 2009/04/18 18:41:52 treectrl Exp $
  */
 
-#include <stdint.h>
 #include "tkTreeCtrl.h"
 
 #ifdef WIN32
@@ -1975,7 +1974,7 @@ Tree_AddItem(
 	int id, isNew;
 
 	id = TreeItem_SetID(tree, item, tree->nextItemId++);
-	hPtr = Tcl_CreateHashEntry(&tree->itemHash, (char *)(intptr_t) id, &isNew);
+	hPtr = Tcl_CreateHashEntry(&tree->itemHash, (char *)(long) id, &isNew);
 	Tcl_SetHashValue(hPtr, item);
 	tree->itemCount++;
 }
@@ -2015,7 +2014,7 @@ Tree_RemoveItem(
 		Tcl_DeleteHashEntry(hPtr);
 
 	hPtr = Tcl_FindHashEntry(&tree->itemHash,
-			(char *)(intptr_t) TreeItem_GetID(tree, item));
+			(char *)(long) TreeItem_GetID(tree, item));
 	Tcl_DeleteHashEntry(hPtr);
 	tree->itemCount--;
 	if (tree->itemCount == 1)
