+++ trunk/src/tk/treectrl/tkTreeItem.c
@@ -1,6 +1,6 @@
 // ======================================================================
-// @version $Revision: 28 $
-// @lastmodified $LastChangedDate: 2011-05-21 14:57:26 +0000 (Sat, 21 May 2011) $
+// @version $Revision: 30 $
+// @lastmodified $LastChangedDate: 2011-05-23 14:49:04 +0000 (Mon, 23 May 2011) $
 // @modifiedby $LastChangedBy$
 // ======================================================================
 
@@ -15,7 +15,6 @@
  * RCS: @(#) $Id: tkTreeItem.c,v 1.107 2009/05/17 18:39:54 treectrl Exp $
  */
 
-#include <stdint.h>
 #include "tkTreeCtrl.h"
 
 typedef struct TreeItem_ TreeItem_;
@@ -2082,7 +2081,7 @@ TreeItemList_FromObj(
 			gotId = TRUE;
 		}
 		if (gotId) {
-			hPtr = Tcl_FindHashEntry(&tree->itemHash, (char *)(intptr_t) id);
+			hPtr = Tcl_FindHashEntry(&tree->itemHash, (char *)(long) id);
 			if (hPtr != NULL) {
 				item = (TreeItem) Tcl_GetHashValue(hPtr);
 			} else {
