+++ trunk/src/tk/treectrl/tkTreeUtils.c
@@ -1,7 +1,7 @@
 // ======================================================================
 // Author : $Author$
-// Version: $Revision: 28 $
-// Date   : $Date: 2011-05-21 14:57:26 +0000 (Sat, 21 May 2011) $
+// Version: $Revision: 30 $
+// Date   : $Date: 2011-05-23 14:49:04 +0000 (Mon, 23 May 2011) $
 // Url    : $URL$
 // ======================================================================
 
@@ -15,7 +15,6 @@
  * RCS: @(#) $Id: tkTreeUtils.c,v 1.75 2008/07/21 18:35:38 treectrl Exp $
  */
 
-#include <stdint.h>
 #include "tkTreeCtrl.h"
 
 #ifdef WIN32
@@ -6642,7 +6641,7 @@ BooleanFlagCO_Set(
 	int flags
 	)
 {
-	int theFlag = (int)(intptr_t) clientData;
+	int theFlag = (long) clientData;
 	int new, *internalPtr;
 
 	if (internalOffset >= 0)
@@ -6672,7 +6671,7 @@ BooleanFlagCO_Get(
 	int internalOffset
 	)
 {
-	int theFlag = (int)(intptr_t) clientData;
+	int theFlag = (long) clientData;
 	int value = *(int *) (recordPtr + internalOffset);
 
 	return Tcl_NewBooleanObj(value & theFlag);
@@ -6686,7 +6685,7 @@ BooleanFlagCO_Restore(
 	char *saveInternalPtr
 	)
 {
-	int theFlag = (int)(intptr_t) clientData;
+	int theFlag = (long) clientData;
 	int value = *(int *) saveInternalPtr;
 
 	if (value & theFlag)
@@ -6718,7 +6717,7 @@ BooleanFlagCO_Init(
 	co->getProc = BooleanFlagCO_Get;
 	co->restoreProc = BooleanFlagCO_Restore;
 	co->freeProc = NULL;
-	co->clientData = (ClientData)(intptr_t) theFlag;
+	co->clientData = (ClientData)(long) theFlag;
 
 	specPtr->clientData = co;
